refactor(core): separate/refactor connector route UTs
