+++ packages/core/src/routes/connector.test.ts
@@ -1,46 +1,17 @@
-/* eslint-disable max-lines, max-nested-callbacks */
-import {
-  ConnectorError,
-  ConnectorErrorCodes,
-  EmailMessageTypes,
-  SmsMessageTypes,
-  ValidateConfig,
-} from '@logto/connector-types';
+import { EmailMessageTypes, ValidateConfig } from '@logto/connector-types';
 import { Connector, ConnectorType } from '@logto/schemas';
 
-import { mockConnectorInstanceList, mockConnectorList } from '@/__mocks__';
+import { mockConnectorInstanceList, mockMetadata, mockConnector } from '@/__mocks__';
 import {
   ConnectorMetadata,
   EmailConnectorInstance,
   SmsConnectorInstance,
 } from '@/connectors/types';
 import RequestError from '@/errors/RequestError';
-import { updateConnector } from '@/queries/connector';
 import assertThat from '@/utils/assert-that';
 import { createRequester } from '@/utils/test-utils';
 
 import connectorRoutes from './connector';
-
-const mockMetadata: ConnectorMetadata = {
-  id: 'id0',
-  target: 'connector_0',
-  type: ConnectorType.Social,
-  platform: null,
-  name: {
-    en: 'Connector',
-    'zh-CN': '连接器',
-  },
-  logo: './logo.png',
-  description: { en: 'Connector', 'zh-CN': '连接器' },
-  readme: 'README.md',
-  configTemplate: 'config-template.md',
-};
-const mockConnector: Connector = {
-  id: 'id0',
-  enabled: true,
-  config: {},
-  createdAt: 1_234_567_890_123,
-};
 
 type ConnectorInstance = {
   connector: Connector;
@@ -49,21 +20,26 @@ type ConnectorInstance = {
   sendMessage?: unknown;
 };
 
-const getConnectorInstanceByIdPlaceHolder = jest.fn() as jest.MockedFunction<
-  (connectorId: string) => Promise<ConnectorInstance>
->;
 const getConnectorInstancesPlaceHolder = jest.fn() as jest.MockedFunction<
   () => Promise<ConnectorInstance[]>
 >;
 
-jest.mock('@/queries/connector', () => ({
-  findAllConnectors: jest.fn(),
-  updateConnector: jest.fn(),
-}));
 jest.mock('@/connectors', () => ({
-  getConnectorInstanceById: async (connectorId: string) =>
-    getConnectorInstanceByIdPlaceHolder(connectorId),
   getConnectorInstances: async () => getConnectorInstancesPlaceHolder(),
+  getConnectorInstanceById: async (connectorId: string) => {
+    const connectorInstances = await getConnectorInstancesPlaceHolder();
+    const connector = connectorInstances.find(({ connector }) => connector.id === connectorId);
+    assertThat(
+      connector,
+      new RequestError({
+        code: 'entity.not_found',
+        connectorId,
+        status: 404,
+      })
+    );
+
+    return connector;
+  },
 }));
 
 describe('connector route', () => {
@@ -107,353 +83,20 @@ describe('connector route', () => {
     });
 
     it('throws when connector can not be found by given connectorId (locally)', async () => {
-      getConnectorInstanceByIdPlaceHolder.mockImplementationOnce(async (_id: string) => {
-        const found = mockConnectorInstanceList.find(
-          (connectorInstance) => connectorInstance.connector.id === 'connector'
-        );
-        assertThat(found, new RequestError({ code: 'entity.not_found', status: 404 }));
-
-        return found;
-      });
+      getConnectorInstancesPlaceHolder.mockResolvedValueOnce(mockConnectorInstanceList.slice(2));
       const response = await connectorRequest.get('/connectors/findConnector').send({});
       expect(response).toHaveProperty('statusCode', 404);
     });
 
     it('throws when connector can not be found by given connectorId (remotely)', async () => {
-      getConnectorInstanceByIdPlaceHolder.mockImplementationOnce(async (id: string) => {
-        const foundConnectorInstance = mockConnectorInstanceList.find(
-          (connectorInstance) => connectorInstance.connector.id === id
-        );
-        assertThat(
-          foundConnectorInstance,
-          new RequestError({ code: 'entity.not_found', status: 404 })
-        );
-
-        const foundConnector = mockConnectorList.find((connector) => connector.id === 'connector0');
-        assertThat(foundConnector, 'entity.not_found');
-
-        return { foundConnector, ...foundConnectorInstance };
-      });
+      getConnectorInstancesPlaceHolder.mockResolvedValueOnce([]);
       const response = await connectorRequest.get('/connectors/id0').send({});
-      expect(response).toHaveProperty('statusCode', 400);
+      expect(response).toHaveProperty('statusCode', 404);
     });
 
     it('shows found connector information', async () => {
-      getConnectorInstanceByIdPlaceHolder.mockImplementationOnce(async (id: string) => {
-        const foundConnectorInstance = mockConnectorInstanceList.find(
-          (connectorInstance) => connectorInstance.connector.id === id
-        );
-        assertThat(
-          foundConnectorInstance,
-          new RequestError({ code: 'entity.not_found', status: 404 })
-        );
-
-        const foundConnector = mockConnectorList.find((connector) => connector.id === id);
-        assertThat(foundConnector, 'entity.not_found');
-
-        return { foundConnector, ...foundConnectorInstance };
-      });
-      const response = await connectorRequest.get('/connectors/id0').send({});
-      expect(response).toHaveProperty('statusCode', 200);
-    });
-  });
-
-  describe('PATCH /connectors/:id/enabled', () => {
-    afterEach(() => {
-      jest.clearAllMocks();
-    });
-
-    it('throws if connector can not be found (locally)', async () => {
-      getConnectorInstanceByIdPlaceHolder.mockImplementationOnce(async (_id: string) => {
-        const found = mockConnectorInstanceList.find(
-          (connectorInstance) => connectorInstance.connector.id === 'connector'
-        );
-        assertThat(found, new RequestError({ code: 'entity.not_found', status: 404 }));
-
-        return found;
-      });
-      const response = await connectorRequest
-        .patch('/connectors/findConnector/enabled')
-        .send({ enabled: true });
-      expect(response).toHaveProperty('statusCode', 404);
-    });
-
-    it('throws if connector can not be found (remotely)', async () => {
-      getConnectorInstanceByIdPlaceHolder.mockImplementationOnce(async (id: string) => {
-        const foundConnectorInstance = mockConnectorInstanceList.find(
-          (connectorInstance) => connectorInstance.connector.id === id
-        );
-        assertThat(
-          foundConnectorInstance,
-          new RequestError({ code: 'entity.not_found', status: 404 })
-        );
-
-        const foundConnector = mockConnectorList.find((connector) => connector.id === 'connector0');
-        assertThat(foundConnector, 'entity.not_found');
-
-        return { foundConnector, ...foundConnectorInstance };
-      });
-      const response = await connectorRequest
-        .patch('/connectors/id0/enabled')
-        .send({ enabled: true });
-      expect(response).toHaveProperty('statusCode', 400);
-    });
-
-    it('enables one of the social connectors (with valid config)', async () => {
-      getConnectorInstanceByIdPlaceHolder.mockImplementationOnce(async (_id: string) => {
-        return {
-          connector: mockConnector,
-          metadata: mockMetadata,
-          validateConfig: jest.fn(),
-        };
-      });
-      const response = await connectorRequest
-        .patch('/connectors/connector_0/enabled')
-        .send({ enabled: true });
-      expect(updateConnector).toHaveBeenCalledWith(
-        expect.objectContaining({
-          where: { id: 'connector_0' },
-          set: { enabled: true },
-        })
-      );
-      expect(response.body).toMatchObject({
-        metadata: mockMetadata,
-      });
-      expect(response).toHaveProperty('statusCode', 200);
-    });
-
-    it('enables one of the social connectors (with invalid config)', async () => {
-      getConnectorInstanceByIdPlaceHolder.mockImplementationOnce(async (_id: string) => {
-        return {
-          connector: mockConnector,
-          metadata: mockMetadata,
-          validateConfig: async () => {
-            throw new ConnectorError(ConnectorErrorCodes.InvalidConfig);
-          },
-        };
-      });
-      const response = await connectorRequest
-        .patch('/connectors/connector_0/enabled')
-        .send({ enabled: true });
-      expect(response).toHaveProperty('statusCode', 500);
-    });
-
-    it('disables one of the social connectors', async () => {
-      getConnectorInstanceByIdPlaceHolder.mockImplementationOnce(async (_id: string) => {
-        return {
-          connector: mockConnector,
-          metadata: mockMetadata,
-          validateConfig: async () => {
-            throw new ConnectorError(ConnectorErrorCodes.InvalidConfig);
-          },
-        };
-      });
-      const response = await connectorRequest
-        .patch('/connectors/connector_0/enabled')
-        .send({ enabled: false });
-      expect(updateConnector).toHaveBeenCalledWith(
-        expect.objectContaining({
-          where: { id: 'connector_0' },
-          set: { enabled: false },
-        })
-      );
-      expect(response.body).toMatchObject({
-        metadata: mockMetadata,
-      });
-      expect(response).toHaveProperty('statusCode', 200);
-    });
-
-    it('enables one of the email/sms connectors (with valid config)', async () => {
       getConnectorInstancesPlaceHolder.mockResolvedValueOnce(mockConnectorInstanceList);
-      const mockedMetadata = {
-        ...mockMetadata,
-        id: 'id1',
-        type: ConnectorType.SMS,
-      };
-      const mockedConnector = {
-        ...mockConnector,
-        id: 'id1',
-        name: 'connector_1',
-        platform: null,
-        type: ConnectorType.SMS,
-        metadata: mockedMetadata,
-      };
-      getConnectorInstanceByIdPlaceHolder.mockImplementationOnce(async (_id: string) => {
-        return {
-          connector: mockedConnector,
-          metadata: mockedMetadata,
-          validateConfig: jest.fn(),
-        };
-      });
-      const response = await connectorRequest
-        .patch('/connectors/id1/enabled')
-        .send({ enabled: true });
-      expect(updateConnector).toHaveBeenNthCalledWith(
-        1,
-        expect.objectContaining({
-          where: { id: 'id1' },
-          set: { enabled: false },
-        })
-      );
-      expect(updateConnector).toHaveBeenNthCalledWith(
-        2,
-        expect.objectContaining({
-          where: { id: 'id5' },
-          set: { enabled: false },
-        })
-      );
-      expect(updateConnector).toHaveBeenNthCalledWith(
-        3,
-        expect.objectContaining({
-          where: { id: 'id1' },
-          set: { enabled: true },
-        })
-      );
-      expect(response.body).toMatchObject({
-        metadata: mockedMetadata,
-      });
-      expect(response).toHaveProperty('statusCode', 200);
-    });
-
-    it('enables one of the email/sms connectors (with invalid config)', async () => {
-      const mockedMetadata = {
-        ...mockMetadata,
-        id: 'connector_1',
-        type: ConnectorType.SMS,
-      };
-      const mockedConnector = {
-        ...mockConnector,
-        id: 'connector_1',
-        name: 'connector_1',
-        platform: null,
-        type: ConnectorType.SMS,
-        metadata: mockedMetadata,
-      };
-      getConnectorInstanceByIdPlaceHolder.mockImplementationOnce(async (_id: string) => {
-        return {
-          connector: mockedConnector,
-          metadata: mockedMetadata,
-          validateConfig: async () => {
-            throw new ConnectorError(ConnectorErrorCodes.InvalidConfig);
-          },
-        };
-      });
-      const response = await connectorRequest
-        .patch('/connectors/connector_1/enabled')
-        .send({ enabled: true });
-      expect(response).toHaveProperty('statusCode', 500);
-    });
-
-    it('disables one of the email/sms connectors', async () => {
-      const mockedMetadata = {
-        ...mockMetadata,
-        id: 'connector_4',
-        type: ConnectorType.Email,
-        platform: null,
-      };
-      const mockedConnector = {
-        ...mockConnector,
-        id: 'connector_4',
-        name: 'connector_4',
-        platform: null,
-        type: ConnectorType.Email,
-        metadata: mockedMetadata,
-      };
-      getConnectorInstanceByIdPlaceHolder.mockImplementationOnce(async (_id: string) => {
-        return {
-          connector: mockedConnector,
-          metadata: mockedMetadata,
-        };
-      });
-      const response = await connectorRequest
-        .patch('/connectors/connector_4/enabled')
-        .send({ enabled: false });
-      expect(updateConnector).toHaveBeenCalledWith(
-        expect.objectContaining({
-          where: { id: 'connector_4' },
-          set: { enabled: false },
-        })
-      );
-      expect(response.body).toMatchObject({
-        metadata: mockedMetadata,
-      });
-      expect(response).toHaveProperty('statusCode', 200);
-    });
-  });
-
-  describe('PATCH /connectors/:id', () => {
-    afterEach(() => {
-      jest.clearAllMocks();
-    });
-
-    it('throws when connector can not be found by given connectorId (locally)', async () => {
-      getConnectorInstanceByIdPlaceHolder.mockImplementationOnce(async (_id: string) => {
-        const found = mockConnectorInstanceList.find(
-          (connectorInstance) => connectorInstance.connector.id === 'connector'
-        );
-        assertThat(found, new RequestError({ code: 'entity.not_found', status: 404 }));
-
-        return found;
-      });
-      const response = await connectorRequest.patch('/connectors/findConnector').send({});
-      expect(response).toHaveProperty('statusCode', 404);
-    });
-
-    it('throws when connector can not be found by given connectorId (remotely)', async () => {
-      getConnectorInstanceByIdPlaceHolder.mockImplementationOnce(async (id: string) => {
-        const foundConnectorInstance = mockConnectorInstanceList.find(
-          (connectorInstance) => connectorInstance.connector.id === id
-        );
-        assertThat(
-          foundConnectorInstance,
-          new RequestError({ code: 'entity.not_found', status: 404 })
-        );
-
-        const foundConnector = mockConnectorList.find((connector) => connector.id === 'id_0');
-        assertThat(foundConnector, 'entity.not_found');
-
-        return { foundConnector, ...foundConnectorInstance };
-      });
-      const response = await connectorRequest.patch('/connectors/id0').send({});
-      expect(response).toHaveProperty('statusCode', 400);
-    });
-
-    it('config validation fails', async () => {
-      getConnectorInstanceByIdPlaceHolder.mockImplementationOnce(async (_id: string) => {
-        return {
-          connector: mockConnector,
-          metadata: mockMetadata,
-          validateConfig: async () => {
-            throw new ConnectorError(ConnectorErrorCodes.InvalidConfig);
-          },
-        };
-      });
-      const response = await connectorRequest
-        .patch('/connectors/connector_0')
-        .send({ config: { cliend_id: 'client_id', client_secret: 'client_secret' } });
-      expect(response).toHaveProperty('statusCode', 500);
-    });
-
-    it('successfully updates connector configs', async () => {
-      getConnectorInstanceByIdPlaceHolder.mockImplementationOnce(async (_id: string) => {
-        return {
-          connector: mockConnector,
-          metadata: mockMetadata,
-          validateConfig: jest.fn(),
-        };
-      });
-      const response = await connectorRequest
-        .patch('/connectors/connector_0')
-        .send({ config: { cliend_id: 'client_id', client_secret: 'client_secret' } });
-      expect(updateConnector).toHaveBeenCalledWith(
-        expect.objectContaining({
-          where: { id: 'connector_0' },
-          set: { config: { cliend_id: 'client_id', client_secret: 'client_secret' } },
-        })
-      );
-      expect(response.body).toMatchObject({
-        metadata: mockMetadata,
-      });
+      const response = await connectorRequest.get('/connectors/id0').send({});
       expect(response).toHaveProperty('statusCode', 200);
     });
   });
@@ -464,18 +107,9 @@ describe('connector route', () => {
     });
 
     it('should get email connector and send message', async () => {
-      const mockedMetadata = {
-        ...mockMetadata,
-        type: ConnectorType.Email,
-      };
-      const mockedConnector = {
-        ...mockConnector,
-        type: ConnectorType.Email,
-        metadata: mockedMetadata,
-      };
       const mockedEmailConnector: EmailConnectorInstance = {
-        connector: mockedConnector,
-        metadata: mockedMetadata,
+        connector: mockConnector,
+        metadata: mockMetadata,
         validateConfig: jest.fn(),
         getConfig: jest.fn(),
         sendMessage: async (
@@ -485,9 +119,7 @@ describe('connector route', () => {
           // eslint-disable-next-line @typescript-eslint/no-empty-function
         ): Promise<any> => {},
       };
-
       getConnectorInstancesPlaceHolder.mockResolvedValueOnce([mockedEmailConnector]);
-
       const sendMessageSpy = jest.spyOn(mockedEmailConnector, 'sendMessage');
       const response = await connectorRequest
         .post('/connectors/test/email')
@@ -500,29 +132,7 @@ describe('connector route', () => {
     });
 
     it('should throw when email connector is not found', async () => {
-      const mockedMetadata = {
-        ...mockMetadata,
-        type: ConnectorType.SMS,
-      };
-      const mockedConnector = {
-        ...mockConnector,
-        type: ConnectorType.SMS,
-      };
-      const mockedSmsConnector: SmsConnectorInstance = {
-        connector: mockedConnector,
-        metadata: mockedMetadata,
-        validateConfig: jest.fn(),
-        getConfig: jest.fn(),
-        sendMessage: async (
-          address: string,
-          type: keyof SmsMessageTypes,
-          _payload: SmsMessageTypes[typeof type]
-          // eslint-disable-next-line @typescript-eslint/no-empty-function
-        ): Promise<any> => {},
-      };
-
-      getConnectorInstancesPlaceHolder.mockResolvedValueOnce([mockedSmsConnector]);
-
+      getConnectorInstancesPlaceHolder.mockResolvedValueOnce([]);
       const response = await connectorRequest
         .post('/connectors/test/email')
         .send({ email: 'test@email.com' });
@@ -540,13 +150,8 @@ describe('connector route', () => {
         ...mockMetadata,
         type: ConnectorType.SMS,
       };
-      const mockedConnector = {
-        ...mockConnector,
-        type: ConnectorType.SMS,
-        metadata: mockedMetadata,
-      };
       const mockedSmsConnectorInstance: SmsConnectorInstance = {
-        connector: mockedConnector,
+        connector: mockConnector,
         metadata: mockedMetadata,
         validateConfig: jest.fn(),
         getConfig: jest.fn(),
@@ -557,9 +162,7 @@ describe('connector route', () => {
           // eslint-disable-next-line @typescript-eslint/no-empty-function
         ): Promise<any> => {},
       };
-
       getConnectorInstancesPlaceHolder.mockResolvedValueOnce([mockedSmsConnectorInstance]);
-
       const sendMessageSpy = jest.spyOn(mockedSmsConnectorInstance, 'sendMessage');
       const response = await connectorRequest
         .post('/connectors/test/sms')
@@ -572,30 +175,7 @@ describe('connector route', () => {
     });
 
     it('should throw when sms connector is not found', async () => {
-      const mockedMetadata = {
-        ...mockMetadata,
-        type: ConnectorType.Email,
-      };
-      const mockedConnector = {
-        ...mockConnector,
-        type: ConnectorType.Email,
-        metadata: mockedMetadata,
-      };
-      const mockedEmailConnectorInstance: EmailConnectorInstance = {
-        connector: mockedConnector,
-        metadata: mockedMetadata,
-        validateConfig: jest.fn(),
-        getConfig: jest.fn(),
-        sendMessage: async (
-          address: string,
-          type: keyof EmailMessageTypes,
-          _payload: EmailMessageTypes[typeof type]
-          // eslint-disable-next-line @typescript-eslint/no-empty-function
-        ): Promise<any> => {},
-      };
-
-      getConnectorInstancesPlaceHolder.mockResolvedValueOnce([mockedEmailConnectorInstance]);
-
+      getConnectorInstancesPlaceHolder.mockResolvedValueOnce([]);
       const response = await connectorRequest
         .post('/connectors/test/sms')
         .send({ phone: '12345678901' });
@@ -603,4 +183,3 @@ describe('connector route', () => {
     });
   });
 });
-/* eslint-enable max-lines, max-nested-callbacks */
